Word Class moved out of Lexicon into its own file. Flag resolve check added to Project load.
